+++ src/Command/Cmdlet/WorkingCopy/AddWorkingItemCmdlet.cs
@@ -2,7 +2,7 @@
 
 namespace AstralKeks.SourceControl.Command
 {
-    [Cmdlet(VerbsCommon.Add, Noun.WorkingItem)]
+    [Cmdlet(VerbsCommon.Add, Noun.SCWorkingItem)]
     public class AddWorkingItemCmdlet : WorkingCopyCmdlet
     {
         protected override void ProcessRecord()