+++ src/Command/Cmdlet/WorkingCopy/PopWorkingCopyCmdlet.cs
@@ -2,7 +2,7 @@
 
 namespace AstralKeks.SourceControl.Command
 {
-    [Cmdlet(VerbsCommon.Pop, Noun.WorkingCopy)]
+    [Cmdlet(VerbsCommon.Pop, Noun.SCWorkingCopy)]
     public class PopWorkingCopyCmdlet : WorkingCopyCmdlet
     {
         protected override void ProcessRecord()