+++ src/Command/Cmdlet/WorkingCopy/PushWorkingCopyCmdlet.cs
@@ -2,7 +2,7 @@
 
 namespace AstralKeks.SourceControl.Command
 {
-    [Cmdlet(VerbsCommon.Push, Noun.WorkingCopy)]
+    [Cmdlet(VerbsCommon.Push, Noun.SCWorkingCopy)]
     public class PushWorkingCopyCmdlet : WorkingCopyCmdlet
     {
         protected override void ProcessRecord()