+++ src/Command/Cmdlet/WorkingCopy/ResetWorkingCopyCmdlet.cs
@@ -2,7 +2,7 @@
 
 namespace AstralKeks.SourceControl.Command
 {
-    [Cmdlet(VerbsCommon.Reset, Noun.WorkingCopy)]
+    [Cmdlet(VerbsCommon.Reset, Noun.SCWorkingCopy)]
     public class ResetWorkingCopyCmdlet : WorkingCopyCmdlet
     {
         protected override void ProcessRecord()