+++ src/Command/Cmdlet/WorkingCopy/ShowWorkingCopyDiffCmdlet.cs
@@ -2,7 +2,7 @@
 
 namespace AstralKeks.SourceControl.Command
 {
-    [Cmdlet(VerbsCommon.Show, Noun.WorkingCopyDiff)]
+    [Cmdlet(VerbsCommon.Show, Noun.SCWorkingCopyDiff)]
     public class ShowWorkingCopyDiffCmdlet : WorkingCopyCmdlet
     {
         protected override void ProcessRecord()