+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+### Fixed
+- Export-SCWorkingCopyDiff
+- Import-SCWorkingCopyDiff
+- Broken autocomplete for shortcuts
+
 ## 0.0.2
 ### Fixed
 - Export-SCWorkingCopyDiff